backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
firebase-admin==6.4.0
google-generativeai==0.4.1
pytest==8.1.1
httpx==0.27.0
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
